backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
chromadb==0.5.5
google-generativeai==0.8.3
python-dotenv
pytest==8.3.3
httpx==0.27.2
python-multipart==0.0.20
pypdf==5.4.0
